Use esm syntax.
closes #2
