+++ lib/index.js
@@ -2,154 +2,6 @@
  * Copyright (c) 2020 Digital Bazaar, Inc. All rights reserved.
  */
 'use strict';
-
-const assert = require('assert-plus');
-const noopLogger = require('./noopLogger');
-const {
-  UnauthorizedClient, InvalidGrant, InvalidRequest, InvalidClient
-} = require('@interop/oauth2-errors');
-
-// Default Access token expiration in seconds (two weeks)
-const DEFAULT_MAX_AGE = 1209600;
-
-/**
- * Constructs a middleware handler for OAuth 2.0 /token exchange requests.
- *
- * @param {object} options - Options hashmap.
- * @param {Function} options.authenticateClient - Async callback that
- *   authenticates the client token.
- * @param {object} [options.defaults={}] - Token issuance defaults.
- * @param {object} [options.logger] - Logger.
- * @param {Function} options.getClient - Async callback to load client by ID.
- * @param {Function} options.issue - Async callback that issues the JWT access
- *   token for the client.
- *
- * @example <caption>Example usage.</caption>
- * tokenExchangeHandler({
- *   defaults: {},
- *   logger: console,
- *   getClient: async ({clientId}) => {
- *     // load client from the db, then:
- *     return {client};
- *   }),
- *   issue: async ({client, scope, body, defaultMaxAge}) => {
- *     // issue JWT access token here
- *     return {accessToken, expiresIn};
- *   })
-  },
- * });
- *
- * @returns {Promise<Function>} Resolves with exchange middleware.
- */
-function tokenExchangeHandler({
-  authenticateClient,
-  defaults = {},
-  getClient,
-  issue,
-  logger = noopLogger,
-}) {
-  assert.func(authenticateClient, 'authenticateClient');
-  assert.func(getClient, 'getClient');
-  assert.func(issue, 'issue');
-
-  return async (req, res) => {
-    try {
-      const {body} = req;
-      const {
-        client_id: clientId,
-        client_secret: clientSecret,
-        grant_type: grantType,
-        scope: scopeRequested
-      } = body;
-
-      _validateParams({clientId, clientSecret, grantType});
-
-      let client;
-      try {
-        client = await getClient({clientId});
-      } catch(cause) {
-        const error = new InvalidRequest({
-          description: 'Could not load client.'
-        });
-        error.cause = cause;
-        throw error;
-      }
-
-      const {authenticated} = await authenticateClient({client, clientSecret});
-      if(!authenticated) {
-        throw new InvalidClient({
-          description: 'The client secret could not be authenticated.'
-        });
-      }
-
-      const defaultMaxAge = defaults.defaultMaxAge || DEFAULT_MAX_AGE;
-
-      const {accessToken, expiresIn, scope} = await issue({
-        client, scopeRequested, body, defaultMaxAge
-      });
-
-      const response = {
-        access_token: accessToken,
-        token_type: 'Bearer',
-        expires_in: expiresIn,
-        scope
-      };
-
-      _respond({response, res});
-    } catch(error) {
-      _error({error, logger, res});
-    }
-  };
-}
-
-function _validateParams({clientId, clientSecret, grantType}) {
-  if(!(clientId && clientSecret)) {
-    throw new UnauthorizedClient({
-      description: 'Missing client credentials at the Token endpoint.'
-    });
-  }
-  if(!grantType) {
-    throw new InvalidRequest({
-      description: 'Missing grant_type parameter.'
-    });
-  }
-  if(grantType !== 'client_credentials') {
-    throw new InvalidGrant({
-      error: 'unsupported_grant_type',
-      description: `Grant type "${grantType}" is not supported.`
-    });
-  }
-}
-
-function _respond({response, res}) {
-  res.set({
-    'Cache-Control': 'no-store',
-    Pragma: 'no-cache'
-  });
-
-  res.status(201).json(response);
-}
-
-function _error({error, logger, res}) {
-  logger.error('Token exchange endpoint error.', {error});
-  const {
-    error: errorId, error_description: description, error_uri: uri
-  } = error;
-
-  const statusCode = error.statusCode || 400;
-  const oauth2ErrorResponse = {
-    error: errorId || 'invalid_request',
-    error_description: description || error.message,
-    error_uri: uri
-  };
-  res.set({
-    'Cache-Control': 'no-store',
-    Pragma: 'no-cache'
-  });
-  return res.status(statusCode).json(oauth2ErrorResponse);
-}
-
-module.exports = {
-  tokenExchangeHandler,
-  _respond
-};
+// translate `main.js` to CommonJS
+require = require('esm')(module);
+module.exports = require('./main.js');
