+++ lib/noopLogger.js
@@ -1,9 +1,7 @@
 /*!
  * Copyright (c) 2020 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
-
-module.exports = new Proxy({}, {
+export default new Proxy({}, {
   get() {
     return () => {};
   }
