+++ package.json
@@ -26,7 +26,7 @@
   "homepage": "https://github.com/digitalbazaar/oauth2-token-endpoint-handler",
   "dependencies": {
     "@interop/oauth2-errors": "^2.0.0",
-    "assert-plus": "^1.0.0"
+    "esm": "^3.2.25"
   },
   "directories": {
     "lib": "./lib"
